modified code to run at python anywhere
The `main.py` document was created to be run locally. Later I created an account at PythonAnywhere to run the code hourly.

In order to do that, it is required to install dependencies directly at PythonAnywhere bash console.
Since Selenium is not fully supported I had to change the code to used pyppeteer and BeautifulSoup instead.
For that it was also needed to download and install chrome via the terminal.